modified return statements to include HttpStatus, status and a Custiom MESSAGE

+++ src/main/java/com/personaproject/ecommercewebapp/controllers/CategoryController.java
@@ -2,35 +2,31 @@ package com.personaproject.ecommercewebapp.controllers;
 
 import com.personaproject.ecommercewebapp.dtos.CategoryDTO;
 import com.personaproject.ecommercewebapp.entity.Category;
-import com.personaproject.ecommercewebapp.enums.StandardServiceTokens;
 import com.personaproject.ecommercewebapp.services.CategoryService;
+import com.personaproject.ecommercewebapp.services.HandleAuthentication;
+import com.personaproject.ecommercewebapp.services.ResponseServices;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
-import java.util.Objects;
 
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/api/category")
 public class CategoryController {
 
-    private final StandardServiceTokens standardServiceTokens;
     private final CategoryService categoryService;
-//    private ResponseMessages responseMessages;
+    private final HandleAuthentication handleAuthentication;
+    private final ResponseServices responseServices;
 
     @PostMapping("/create")
-    public String createNewCategory(@RequestBody CategoryDTO categoryDTO,
+    public Object createNewCategory(@RequestBody CategoryDTO categoryDTO,
                                     @RequestHeader String serviceToken,
-                                    @RequestHeader String authToken) throws IllegalAccessException {
-        System.out.println("\t\t |");
-        System.out.println("\t\t |");
-
-        if (!Objects.equals(serviceToken, standardServiceTokens.categoryToken)
-                || !Objects.equals(authToken, standardServiceTokens.authToken)) {
-            throw new IllegalAccessException("Wrong Service and/or authentication Token");
-        }
-        return (categoryService.createCategory(categoryDTO));
+                                    @RequestHeader String authToken) {
+        if (handleAuthentication.authenticateCategoryCreation(authToken, serviceToken))
+            return categoryService.createCategory(categoryDTO);
+        return responseServices.apiResponse(HttpStatus.BAD_REQUEST, false, "Token(s) cannot be validated");
     }
 
     @GetMapping("/get_all")
@@ -39,13 +35,22 @@ public class CategoryController {
     }
 
     @PostMapping("/update/{categoryId}")
-    public String updateCategory(@PathVariable Long categoryId,
+    public Object updateCategory(@PathVariable Long categoryId,
                                  @RequestBody CategoryDTO categoryDTO) {
         return categoryService.updateCategory(categoryId, categoryDTO);
     }
 
     @PostMapping("/delete/{categoryId}")
-    public String deleteCategory(@PathVariable Long categoryId) {
-        return (categoryService.removeCategory(categoryId));
+    public Object deleteCategory(@PathVariable Long categoryId,
+                                 @RequestHeader String serviceToken,
+                                 @RequestHeader String authToken) {
+        try {
+            if (handleAuthentication.authenticateCategoryCreation(authToken, serviceToken))
+                return (categoryService.removeCategory(categoryId));
+            return responseServices.apiResponse(HttpStatus.BAD_REQUEST,
+                    false, "Category with ID " + categoryId + " Cannot be deleted");
+        } catch (Exception e) {
+            return ("Token(s) cannot be verified");
+        }
     }
 }
